Add autoload-dev, remove workarounds in unit tests
