+++ src/Calcine/Model/ParseException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Calcine\Model;
+
+final class ParseException extends \Exception
+{
+}
